Moved ActorView type and processing into commands
