+++ commands/actor.go
@@ -3,12 +3,43 @@ package commands
 import (
 	"encoding/json"
 	"io"
+	"reflect"
+	"strings"
 
-	"github.com/filecoin-project/go-filecoin/plumbing/actr"
+	"github.com/filecoin-project/go-filecoin/actor"
+	"github.com/filecoin-project/go-filecoin/actor/builtin/account"
+	"github.com/filecoin-project/go-filecoin/actor/builtin/miner"
+	"github.com/filecoin-project/go-filecoin/actor/builtin/paymentbroker"
+	"github.com/filecoin-project/go-filecoin/actor/builtin/storagemarket"
+	"github.com/filecoin-project/go-filecoin/exec"
+	"github.com/filecoin-project/go-filecoin/types"
 
 	"gx/ipfs/QmQtQrtNioesAWtrx8csBvfY37gTe94d6wQ3VikZUjxD39/go-ipfs-cmds"
+	"gx/ipfs/QmR8BauakNcBa3RbE4nbQu76PDiJgoQgz8AJdhJuiU4TAw/go-cid"
 	"gx/ipfs/Qmde5VP1qUkyQXKCfmEUA7bP64V2HAptbJ7phuPp7jXWwg/go-ipfs-cmdkit"
 )
+
+// ActorView represents a generic way to represent details about any actor to the user.
+type ActorView struct {
+	ActorType string          `json:"actorType"`
+	Address   string          `json:"address"`
+	Code      cid.Cid         `json:"code,omitempty"`
+	Nonce     uint64          `json:"nonce"`
+	Balance   *types.AttoFIL  `json:"balance"`
+	Exports   readableExports `json:"exports"`
+	Head      cid.Cid         `json:"head,omitempty"`
+}
+
+// readableFunctionSignature is a representation of an actors function signature,
+// such that it can be shown to the user.
+type readableFunctionSignature struct {
+	Params []string
+	Return []string
+}
+
+// readableExports is a representation of exports (map of method names to signatures),
+// such that it can be shown to the user.
+type readableExports map[string]*readableFunctionSignature
 
 var actorCmd = &cmds.Command{
 	Helptext: cmdkit.HelpText{
@@ -21,21 +52,40 @@ var actorCmd = &cmds.Command{
 
 var actorLsCmd = &cmds.Command{
 	Run: func(req *cmds.Request, re cmds.ResponseEmitter, env cmds.Environment) error {
-		actors, err := GetPorcelainAPI(env).ActorLs(req.Context)
+		addrs, actors, err := GetPorcelainAPI(env).ActorLs(req.Context)
 		if err != nil {
 			return err
 		}
 
-		for _, actor := range actors {
-			if err := re.Emit(actor); err != nil {
+		for i, a := range actors {
+			var result *ActorView
+
+			switch {
+			case a.Empty(): // empty (balance only) actors have no Code.
+				result = makeActorView(a, addrs[i], nil)
+			case a.Code.Equals(types.AccountActorCodeCid):
+				result = makeActorView(a, addrs[i], &account.Actor{})
+			case a.Code.Equals(types.StorageMarketActorCodeCid):
+				result = makeActorView(a, addrs[i], &storagemarket.Actor{})
+			case a.Code.Equals(types.PaymentBrokerActorCodeCid):
+				result = makeActorView(a, addrs[i], &paymentbroker.Actor{})
+			case a.Code.Equals(types.MinerActorCodeCid):
+				result = makeActorView(a, addrs[i], &miner.Actor{})
+			case a.Code.Equals(types.BootstrapMinerActorCodeCid):
+				result = makeActorView(a, addrs[i], &miner.Actor{})
+			default:
+				result = makeActorView(a, addrs[i], nil)
+			}
+
+			if err := re.Emit(result); err != nil {
 				return err
 			}
 		}
 		return nil
 	},
-	Type: &actr.ActorView{},
+	Type: &ActorView{},
 	Encoders: cmds.EncoderMap{
-		cmds.JSON: cmds.MakeTypedEncoder(func(req *cmds.Request, w io.Writer, a *actr.ActorView) error {
+		cmds.JSON: cmds.MakeTypedEncoder(func(req *cmds.Request, w io.Writer, a *ActorView) error {
 			marshaled, err := json.Marshal(a)
 			if err != nil {
 				return err
@@ -48,4 +98,54 @@ var actorLsCmd = &cmds.Command{
 			return err
 		}),
 	},
+}
+
+func makeActorView(act *actor.Actor, addr string, actType exec.ExecutableActor) *ActorView {
+	var actorType string
+	var exports readableExports
+	if actType == nil {
+		actorType = "UnknownActor"
+	} else {
+		actorType = getActorType(actType)
+		exports = presentExports(actType.Exports())
+	}
+
+	return &ActorView{
+		ActorType: actorType,
+		Address:   addr,
+		Code:      act.Code,
+		Nonce:     uint64(act.Nonce),
+		Balance:   act.Balance,
+		Exports:   exports,
+		Head:      act.Head,
+	}
+}
+
+func makeReadable(f *exec.FunctionSignature) *readableFunctionSignature {
+	rfs := &readableFunctionSignature{
+		Params: make([]string, len(f.Params)),
+		Return: make([]string, len(f.Return)),
+	}
+	for i, p := range f.Params {
+		rfs.Params[i] = p.String()
+	}
+	for i, r := range f.Return {
+		rfs.Return[i] = r.String()
+	}
+	return rfs
+}
+
+func presentExports(e exec.Exports) readableExports {
+	rdx := make(readableExports)
+	for k, v := range e {
+		rdx[k] = makeReadable(v)
+	}
+	return rdx
+}
+
+func getActorType(actType exec.ExecutableActor) string {
+	t := reflect.TypeOf(actType).Elem()
+	prefixes := strings.Split(t.PkgPath(), "/")
+
+	return strings.Title(prefixes[len(prefixes)-1]) + t.Name()
 }
